@@ -3,8 +3,8 @@ package net.focik.homeoffice.library.api;
 import lombok.AllArgsConstructor;
 import net.focik.homeoffice.library.api.dto.BookApiDto;
 import net.focik.homeoffice.library.api.mapper.ApiBookMapper;
+import net.focik.homeoffice.library.domain.LibraryFacade;
 import net.focik.homeoffice.library.domain.model.Book;
-import net.focik.homeoffice.library.domain.model.WebSite;
 import net.focik.homeoffice.library.domain.port.primary.DeleteBookUseCase;
 import net.focik.homeoffice.library.domain.port.primary.FindBookUseCase;
 import net.focik.homeoffice.library.domain.port.primary.SaveBookUseCase;
@@ -29,50 +29,60 @@ public class BookController {
     private final ModelMapper mapper;
     private final ApiBookMapper bookMapper;
 
+    private final  LibraryFacade testFacade;
+
+
+    @GetMapping("/test")
+    public void test()    {
+        testFacade.findNewBooksInSeriesScheduler();
+    }
     @GetMapping("/{id}")
-    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ')")
+    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ') or hasRole('ROLE_ADMIN')")
     ResponseEntity<BookApiDto> getById(@PathVariable int id) {
         Book book = findBookUseCase.findBook(id);
         return new ResponseEntity<>(bookMapper.toDto(book), HttpStatus.OK);
     }
 
     @GetMapping
-    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ')")
+    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ') or hasRole('ROLE_ADMIN')")
     ResponseEntity<List<BookApiDto>> getAllBooks() {
         List<Book> allBooks = findBookUseCase.findAllBooks();
         return new ResponseEntity<>(allBooks.stream().map(bookMapper::toDto).collect(Collectors.toList()), HttpStatus.OK);
     }
 
     @GetMapping("/url")
-    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ')")
-    ResponseEntity<BookApiDto> getBookByUrl(@RequestParam(name = "site") WebSite webSite, @RequestParam(name = "url") String url) {
-        Book bookDtoByUrl = findBookUseCase.findBookByUrl(webSite, url);
+    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ') or hasRole('ROLE_ADMIN')")
+    ResponseEntity<BookApiDto> getBookByUrl(@RequestParam(name = "url") String url) {
+        Book bookDtoByUrl = findBookUseCase.findBookByUrl(url);
         BookApiDto dto = bookMapper.toDto(bookDtoByUrl);
         return new ResponseEntity<>(dto, HttpStatus.OK);
     }
 
     @GetMapping("/series/{id}")
-    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ')")
+    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ') or hasRole('ROLE_ADMIN')")
     ResponseEntity<List<BookApiDto>> getAllBooksInSeries(@PathVariable int id) {
         List<BookApiDto> existedBooks = findBookUseCase.findAllBooksInSeries(id).stream()
                 .map(bookMapper::toDto)
-                .collect(Collectors.toList());
-        List<BookApiDto> newBooks = findBookUseCase.findNewBooksInSeries(id).stream()
-                .map(bookDto -> mapper.map(bookDto, BookApiDto.class))
-                .collect(Collectors.toList());
-
-        for (BookApiDto dto : newBooks) {
-            if (existedBooks.stream().noneMatch(bookApiDto -> dto.getTitle().equals(bookApiDto.getTitle()))) {
-                existedBooks.add(dto);
-            }
-        }
+                .toList();
         return new ResponseEntity<>(existedBooks.stream()
                 .sorted(Comparator.comparing(BookApiDto::getBookInSeriesNo))
                 .collect(Collectors.toList()), HttpStatus.OK);
     }
 
+    @GetMapping("/series/new/{id}")
+    @PreAuthorize("hasAnyAuthority('LIBRARY_READ_ALL','LIBRARY_READ') or hasRole('ROLE_ADMIN')")
+    ResponseEntity<List<BookApiDto>> getNewBooksInSeries(@PathVariable int id, @RequestParam( name = "url") String url) {
+        List<BookApiDto> newBooks = findBookUseCase.findNewBooksInSeriesByUrl(id, url).stream()
+                .map(bookDto -> mapper.map(bookDto, BookApiDto.class))
+                .toList();
+
+        return new ResponseEntity<>(newBooks.stream()
+                .sorted(Comparator.comparing(BookApiDto::getBookInSeriesNo))
+                .collect(Collectors.toList()), HttpStatus.OK);
+    }
+
     @PostMapping
-    @PreAuthorize("hasAnyAuthority('LIBRARY_WRITE_ALL','LIBRARY_WRITE')")
+    @PreAuthorize("hasAnyAuthority('LIBRARY_WRITE_ALL','LIBRARY_WRITE') or hasRole('ROLE_ADMIN')")
     public ResponseEntity<BookApiDto> addBook(@RequestBody BookApiDto bookDto) {
         Book bookToAdd = bookMapper.toDomain(bookDto);
         Book bookAdded = saveBookUseCase.addBook(bookToAdd);
@@ -80,7 +90,7 @@ public class BookController {
     }
 
     @PutMapping()
-    @PreAuthorize("hasAnyAuthority('LIBRARY_WRITE_ALL','LIBRARY_WRITE')")
+    @PreAuthorize("hasAnyAuthority('LIBRARY_WRITE_ALL','LIBRARY_WRITE') or hasRole('ROLE_ADMIN')")
     public ResponseEntity<BookApiDto> editBook(@RequestBody BookApiDto bookDto) {
         Book bookToUpdate = bookMapper.toDomain(bookDto);
         Book updateBook = saveBookUseCase.updateBook(bookToUpdate);
@@ -88,7 +98,7 @@ public class BookController {
     }
 
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasAnyAuthority('LIBRARY_DELETE_ALL','LIBRARY_DELETE')")
+    @PreAuthorize("hasAnyAuthority('LIBRARY_DELETE_ALL','LIBRARY_DELETE') or hasRole('ROLE_ADMIN')")
     public void deleteBook(@PathVariable Integer id) {
         deleteBookUseCase.deleteBook(id);
     }
