@@ -1,9 +1,10 @@
 package net.focik.homeoffice.finance.api.dto;
 
-import lombok.Builder;
-import lombok.Getter;
-import lombok.ToString;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.*;
 
+@AllArgsConstructor
+@NoArgsConstructor
 @Builder
 @Getter
 @ToString
@@ -19,5 +20,6 @@ public class PurchaseDto {
     private String paymentDate;
     private String otherInfo;
     private PaymentStatusDto paymentStatus;
+    @JsonProperty("installment")
     private boolean isInstallment;
 }
