@@ -18,7 +18,7 @@ public class Book {
     private String title;
     private String description;
     private String cover;
-    private Integer bookInSeriesNo;
+    private String bookInSeriesNo;
 
     @Override
     public boolean equals(Object o) {
