@@ -2,7 +2,7 @@ package net.focik.homeoffice.library.domain.model;
 
 
 public enum WebSite {
-    UPOLUJ_EBOOKA("upolujebooka.pl"), EMPIK("empik.com");
+    UPOLUJ_EBOOKA("upolujebooka.pl"), LEGIMI("legimi.pl"), LUBIMY_CZYTAC("lubimyczytac.pl"), CHAT_GPT("open.ai");
 
     private final String url;
 
