@@ -1,10 +1,12 @@
 package net.focik.homeoffice.userservice.domain;
 
+import jakarta.persistence.*;
 import lombok.*;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
-import javax.persistence.*;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 
 @Entity
 @NoArgsConstructor
@@ -12,7 +14,7 @@ import java.util.List;
 @Getter
 @Setter
 @Table(name = "users")
-public class AppUser {
+public class AppUser implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -28,22 +30,17 @@ public class AppUser {
     private Date joinDate;
     private Date lastLoginDate;
     private Date lastLoginDateDisplay;
+
+    @Override
+    public String toString() {
+        return  username;
+    }
+
     @Column(name = "is_not_locked")
     private boolean notLocked;
-//    private boolean isNotLocked;
     @OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL, orphanRemoval = true)
     @JoinColumn(name = "user_id")
-//    @JoinTable(
-//            name = "users_privileges",
-//            joinColumns = @JoinColumn(
-//                    name = "user_id", referencedColumnName = "id"),
-//            inverseJoinColumns = @JoinColumn(
-//                    name = "privilege_id", referencedColumnName = "id"))
     private List<Privilege> privileges;
-
-    public AppUser(Long id) {
-        this.id = id;
-    }
 
 
     public void addPrivilege(Privilege privilege){
@@ -52,5 +49,51 @@ public class AppUser {
 
     public void deletePrivilege(Privilege privilege) {
         privileges.remove(privilege);
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return getGrantedAuthorities(getPrivileges(privileges));
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return notLocked;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+
+    public boolean isAdmin(){
+        Optional<Privilege> isAdmin = getPrivileges().stream()
+                .filter(privilege -> "ROLE_ADMIN".equals(privilege.getRole().getName()))
+                .findFirst();
+        return isAdmin.isPresent();
+    }
+    private List<String> getPrivileges(Collection<Privilege> privileges) {
+
+        List<String> privilegesList = new ArrayList<>();
+        for (Privilege privilege : privileges) {
+            privilegesList.add(privilege.getRole().getName());
+            privilegesList.add(privilege.getFullReadName());
+            privilegesList.add(privilege.getFullWriteName());
+            privilegesList.add(privilege.getFullDeleteName());
+        }
+        return privilegesList;
+    }
+    private List<GrantedAuthority> getGrantedAuthorities(List<String> privileges) {
+        List<GrantedAuthority> authorities = new ArrayList<>();
+        for (String privilege : privileges) {
+            authorities.add(new SimpleGrantedAuthority(privilege));
+        }
+        return authorities;
     }
 }
