@@ -1,13 +1,9 @@
 package net.focik.homeoffice.userservice.domain;
 
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.Reader;
 
 @Entity
 @NoArgsConstructor
@@ -19,29 +15,5 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     private String name;
-//    private PrivilegeType read;
-//    private PrivilegeType write;
-//    private PrivilegeType delete;
-//    @ManyToMany(mappedBy = "roles")
-//    private Collection<AppUser> users;
-
-//    @ManyToMany(fetch = FetchType.LAZY)
-//    @JoinTable(
-//            name = "roles_privileges",
-//            joinColumns = @JoinColumn(
-//                    name = "role_id", referencedColumnName = "id"),
-//            inverseJoinColumns = @JoinColumn(
-//                    name = "privilege_id", referencedColumnName = "id"))
-
-//        @OneToMany(fetch = FetchType.LAZY)
-//    @JoinTable(
-//            name = "roles_privileges",
-//            joinColumns = @JoinColumn(
-//                    name = "role_id", referencedColumnName = "id"),
-//            inverseJoinColumns = @JoinColumn(
-//                    name = "privilege_id", referencedColumnName = "id"))
-//    private Collection<Privilege> privileges;
-
 }
